Refactored the postRequest function, unfortunately there is a status code: 500.
