+++ EulerityApp/Model/ImageResponse.swift
@@ -1,0 +1,12 @@
+//
+//  ImageResponse.swift
+//  EulerityApp
+//
+//  Created by Essa Aldo on 5/10/21.
+//
+
+import Foundation
+
+struct ImageResponse: Decodable {
+    let url: String
+}
